Implemented rest of InvitationsScreenComponent

+++ frontend/composeApp/src/commonMain/kotlin/pw/edu/pl/pap/api/data/TemporaryMembershipApi.kt
@@ -7,10 +7,10 @@ import pw.edu.pl.pap.data.databaseAssociatedData.Invitation
 interface TemporaryMembershipApi {
 
     @POST("tempMembership/accept/{id}")
-    suspend fun accept(@Path("id")  id: Int): HttpResponse
+    suspend fun accept(@Path("id")  id: Long): HttpResponse
 
     @DELETE("tempMembership/delete/{id}")
-    suspend fun decline(@Path("id")  id: Int): HttpResponse
+    suspend fun decline(@Path("id")  id: Long): HttpResponse
 
     @GET("tempMembership/invitations")
     suspend fun getReceived(): List<Invitation>
